C tests: deduplicate some test mains
Deduplicate most more complex mains.
