+++ test/C/common.h
@@ -117,4 +117,60 @@ test_standard_read
   yaep_free_grammar (g);
 }
 
+static void
+test_complex_parse
+  (int one_parse, int ambiguous, int print_cost, int recovery_match,
+   int argc, char **argv)
+{
+  struct grammar *g;
+  struct yaep_tree_node *root;
+  int ambiguous_p;
+
+  if ((g = yaep_create_grammar ()) == NULL)
+    {
+      fprintf (stderr, "yaep_create_grammar: No memory\n");
+      exit (1);
+    }
+  yaep_set_one_parse_flag (g, one_parse);
+  if (print_cost)
+    {
+      yaep_set_cost_flag (g, 1);
+    }
+  if (argc > 1)
+    yaep_set_lookahead_level (g, atoi (argv [1]));
+  if (argc > 2)
+    yaep_set_debug_level (g, atoi (argv [2]));
+  else
+    yaep_set_debug_level (g, 3);
+  if (argc > 3)
+    yaep_set_error_recovery_flag (g, atoi (argv [3]));
+  if (argc > 4)
+    yaep_set_one_parse_flag (g, atoi (argv [4]));
+  if (recovery_match)
+    {
+      yaep_set_recovery_match (g, recovery_match);
+    }
+  if (yaep_parse_grammar (g, 1, description) != 0)
+    {
+      fprintf (stderr, "%s\n", yaep_error_message (g));
+      exit (1);
+    }
+  if (yaep_parse (g, test_read_token, test_syntax_error, test_parse_alloc,
+                  test_parse_free, &root, &ambiguous_p))
+    {
+      fprintf (stderr, "yaep parse: %s\n", yaep_error_message (g));
+      exit (1);
+    }
+  if (ambiguous != ambiguous_p)
+    {
+      fprintf (stderr, "Grammar should be %sambiguous\n", ambiguous ? "" : "un");
+      exit (1);
+    }
+  if (print_cost)
+    {
+      fprintf (stderr, "cost = %d\n", root->val.anode.cost);
+    }
+  yaep_free_grammar (g);
+}
+
 #endif
